commit: feat: error component and fetch error handling
