+++ src/components/app-main/app-main.jsx
@@ -4,7 +4,7 @@ import styles from './app-main.module.css';
 import BurgerIngredients from '../burger-ingredients/burger-ingredients';
 import BurgerConstructor from '../burger-constructor/burger-constructor';
 
-import {testData as burgerData} from '../../data/mockBurgerData'
+import {testData as burgerData} from '../../data/test-burger-data'
 
 function AppMain({ ingredientsData }) {
