+++ src/components/app-main/app-main.module.css
@@ -2,4 +2,5 @@
   display: flex;
   justify-content: center;
   min-height: 0;
+  flex-grow: 1;
 }
