+++ src/components/error-message/error-message.module.css
@@ -1,0 +1,7 @@
+.container {
+  display: flex;
+  flex-grow: 1;
+  justify-content: center;
+  align-items: center;
+  text-align: center
+}
