+++ src/components/modal-overlay/modal-overlay.jsx
@@ -7,17 +7,6 @@ function ModalOverlay({ children, setter }) {
 
   const modalPortal = useRef(document.getElementById('modal-windows'));
 
-  const closeWithEsc = (e) => {
-    if (e.key === 'Escape') {
-      setter(false);
-    }
-  }
-
-  useEffect(() => {
-    window.addEventListener('keydown', closeWithEsc);
-    return () => window.removeEventListener('keydown', closeWithEsc);
-  }, []);
-
   const handleClick = (e) => {
     if (e.target === e.currentTarget) {
       setter(false)
